requirements.txt
openai
groq
python-dotenv
requests
gtts
# Bibliotecas de processamento de linguagem natural
nltk
# ElevenLabs (voz via API)
elevenlabs

# Coqui TTS
TTS==0.22.0
transformers==4.33.0
torch==2.1.0
torchaudio==2.1.0
numpy
scipy
soundfile

# Áudio e sistema
ffmpeg-python

# Outros úteis no projeto
tqdm
PyYAML
